@@ -6,35 +6,52 @@
 // update, you can easily restore the functionality you've added.
 // -----------------------------------------------------------------------------
 
-const path = require('path');
-const crypto = require('crypto');
-const fs = require('fs-extra');
-const chalk = require('chalk');
-const CopyPlugin = require('copy-webpack-plugin');
-const WebpackOnBuildPlugin = require('on-build-webpack');
-const tar = require('tar');
-const glob = require('glob');
-const execSync = require('child_process').execSync;
+/* eslint-disable no-console */
+
+import { execSync } from 'node:child_process';
+import crypto from 'node:crypto';
+import { builtinModules } from 'node:module';
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
+import CopyPlugin from 'copy-webpack-plugin';
+import allPossibleCategories from '@joplin/lib/pluginCategories.json' assert { type: 'json' };
+import chalk from 'chalk';
+import fs from 'fs-extra';
+import * as glob from 'glob';
+import path from 'path';
+import * as tar from 'tar';
+import userConfigFile from './plugin.config.json' assert { type: 'json' };
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 const rootDir = path.resolve(__dirname);
-const userConfigFilename = './plugin.config.json';
-const userConfigPath = path.resolve(rootDir, userConfigFilename);
 const distDir = path.resolve(rootDir, 'dist');
 const srcDir = path.resolve(rootDir, 'src');
 const publishDir = path.resolve(rootDir, 'publish');
 
-const userConfig = Object.assign({}, {
-	extraScripts: [],
-}, fs.pathExistsSync(userConfigPath) ? require(userConfigFilename) : {});
+const userConfig = { extraScripts: [], ...userConfigFile };
 
 const manifestPath = `${srcDir}/manifest.json`;
 const packageJsonPath = `${rootDir}/package.json`;
+const allPossibleScreenshotsType = ['jpg', 'jpeg', 'png', 'gif', 'webp'];
 const manifest = readManifest(manifestPath);
 const pluginArchiveFilePath = path.resolve(publishDir, `${manifest.id}.jpl`);
 const pluginInfoFilePath = path.resolve(publishDir, `${manifest.id}.json`);
 
+// Webpack5 doesn't polyfill by default and displays a warning when attempting to require() built-in
+// node modules. Set these to false to prevent Webpack from warning about not polyfilling these modules.
+// We don't need to polyfill because the plugins run in Electron's Node environment.
+const moduleFallback = {};
+for (const moduleName of builtinModules) {
+	moduleFallback[moduleName] = false;
+}
+
+const getPackageJson = () => {
+	return JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
+};
+
 function validatePackageJson() {
-	const content = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
+	const content = getPackageJson();
 	if (!content.name || content.name.indexOf('joplin-plugin-') !== 0) {
 		console.warn(chalk.yellow(`WARNING: To publish the plugin, the package name should start with "joplin-plugin-" (found "${content.name}") in ${packageJsonPath}`));
 	}
@@ -67,15 +84,48 @@ function currentGitInfo() {
 	}
 }
 
+function validateCategories(categories) {
+	if (!categories) return null;
+	if ((categories.length !== new Set(categories).size)) throw new Error('Repeated categories are not allowed');
+	// eslint-disable-next-line github/array-foreach -- Old code before rule was applied
+	categories.forEach(category => {
+		if (!allPossibleCategories.map(category => { return category.name; }).includes(category)) throw new Error(`${category} is not a valid category. Please make sure that the category name is lowercase. Valid categories are: \n${allPossibleCategories.map(category => { return category.name; })}\n`);
+	});
+}
+
+function validateScreenshots(screenshots) {
+	if (!screenshots) return null;
+	for (const screenshot of screenshots) {
+		if (!screenshot.src) throw new Error('You must specify a src for each screenshot');
+
+		// Avoid attempting to download and verify URL screenshots.
+		if (screenshot.src.startsWith('https://') || screenshot.src.startsWith('http://')) {
+			continue;
+		}
+
+		const screenshotType = screenshot.src.split('.').pop();
+		if (!allPossibleScreenshotsType.includes(screenshotType)) throw new Error(`${screenshotType} is not a valid screenshot type. Valid types are: \n${allPossibleScreenshotsType}\n`);
+
+		const screenshotPath = path.resolve(rootDir, screenshot.src);
+
+		// Max file size is 1MB
+		const fileMaxSize = 1024;
+		const fileSize = fs.statSync(screenshotPath).size / 1024;
+		if (fileSize > fileMaxSize) throw new Error(`Max screenshot file size is ${fileMaxSize}KB. ${screenshotPath} is ${fileSize}KB`);
+	}
+}
+
 function readManifest(manifestPath) {
 	const content = fs.readFileSync(manifestPath, 'utf8');
 	const output = JSON.parse(content);
 	if (!output.id) throw new Error(`Manifest plugin ID is not set in ${manifestPath}`);
+	validateCategories(output.categories);
+	validateScreenshots(output.screenshots);
 	return output;
 }
 
 function createPluginArchive(sourceDir, destPath) {
-	const distFiles = glob.sync(`${sourceDir}/**/*`, { nodir: true })
+	const distFiles = glob.sync(`${sourceDir}/**/*`, { nodir: true, windowsPathsNoEscape: true })
 		.map(f => f.substr(sourceDir.length + 1));
 
 	if (!distFiles.length) throw new Error('Plugin archive was not created because the "dist" directory is empty');
@@ -89,18 +139,22 @@ function createPluginArchive(sourceDir, destPath) {
 			cwd: sourceDir,
 			sync: true,
 		},
-		distFiles
+		distFiles,
 	);
 
 	console.info(chalk.cyan(`Plugin archive has been created in ${destPath}`));
 }
+
+const writeManifest = (manifestPath, content) => {
+	fs.writeFileSync(manifestPath, JSON.stringify(content, null, '\t'), 'utf8');
+};
 
 function createPluginInfo(manifestPath, destPath, jplFilePath) {
 	const contentText = fs.readFileSync(manifestPath, 'utf8');
 	const content = JSON.parse(contentText);
 	content._publish_hash = `sha256:${fileSha256(jplFilePath)}`;
 	content._publish_commit = currentGitInfo();
-	fs.writeFileSync(destPath, JSON.stringify(content, null, '\t'), 'utf8');
+	writeManifest(destPath, content);
 }
 
 function onBuildCompleted() {
@@ -129,15 +183,15 @@ const baseConfig = {
 	},
 };
 
-const pluginConfig = Object.assign({}, baseConfig, {
-	entry: './src/index.ts',
+const pluginConfig = { ...baseConfig, entry: './src/index.ts',
 	resolve: {
 		alias: {
 			api: path.resolve(__dirname, 'api'),
 		},
+		fallback: moduleFallback,
 		// JSON files can also be required from scripts so we include this.
 		// https://github.com/joplin/plugin-bibtex/pull/2
-		extensions: ['.tsx', '.ts', '.js', '.json'],
+		extensions: ['.js', '.tsx', '.ts', '.json'],
 	},
 	output: {
 		filename: 'index.js',
@@ -161,26 +215,31 @@ const pluginConfig = Object.assign({}, baseConfig, {
 				},
 			],
 		}),
-	],
-});
+	] };
 
-const extraScriptConfig = Object.assign({}, baseConfig, {
-	resolve: {
-		alias: {
-			api: path.resolve(__dirname, 'api'),
-		},
-		extensions: ['.tsx', '.ts', '.js', '.json'],
+const extraScriptConfig = { ...baseConfig, resolve: {
+	alias: {
+		api: path.resolve(__dirname, 'api'),
 	},
-});
+	fallback: moduleFallback,
+	extensions: ['.js', '.tsx', '.ts', '.json'],
+} };
 
 const createArchiveConfig = {
 	stats: 'errors-only',
 	entry: './dist/index.js',
+	resolve: {
+		fallback: moduleFallback,
+	},
 	output: {
 		filename: 'index.js',
 		path: publishDir,
 	},
-	plugins: [new WebpackOnBuildPlugin(onBuildCompleted)],
+	plugins: [{
+		apply(compiler) {
+			compiler.hooks.done.tap('archiveOnBuildListener', onBuildCompleted);
+		},
+	}],
 };
 
 function resolveExtraScriptPath(name) {
@@ -212,20 +271,41 @@ function buildExtraScriptConfigs(userConfig) {
 
 	for (const scriptName of userConfig.extraScripts) {
 		const scriptPaths = resolveExtraScriptPath(scriptName);
-		output.push(Object.assign({}, extraScriptConfig, {
-			entry: scriptPaths.entry,
-			output: scriptPaths.output,
-		}));
+		output.push({ ...extraScriptConfig, entry: scriptPaths.entry,
+			output: scriptPaths.output });
 	}
 
 	return output;
 }
 
-function main(processArgv) {
-	const yargs = require('yargs/yargs');
-	const argv = yargs(processArgv).argv;
+const increaseVersion = version => {
+	try {
+		const s = version.split('.');
+		const d = Number(s[s.length - 1]) + 1;
+		s[s.length - 1] = `${d}`;
+		return s.join('.');
+	} catch (error) {
+		error.message = `Could not parse version number: ${version}: ${error.message}`;
+		throw error;
+	}
+};
 
-	const configName = argv['joplin-plugin-config'];
+const updateVersion = () => {
+	const packageJson = getPackageJson();
+	packageJson.version = increaseVersion(packageJson.version);
+	fs.writeFileSync(packageJsonPath, `${JSON.stringify(packageJson, null, 2)}\n`, 'utf8');
+
+	const manifest = readManifest(manifestPath);
+	manifest.version = increaseVersion(manifest.version);
+	writeManifest(manifestPath, manifest);
+
+	if (packageJson.version !== manifest.version) {
+		console.warn(chalk.yellow(`Version numbers have been updated but they do not match: package.json (${packageJson.version}), manifest.json (${manifest.version}). Set them to the required values to get them in sync.`));
+	}
+};
+
+function main(environ) {
+	const configName = environ['joplin-plugin-config'];
 	if (!configName) throw new Error('A config file must be specified via the --joplin-plugin-config flag');
 
 	// Webpack configurations run in parallel, while we need them to run in
@@ -260,22 +340,30 @@ function main(processArgv) {
 		fs.mkdirpSync(publishDir);
 	}
 
+	if (configName === 'updateVersion') {
+		updateVersion();
+		return [];
+	}
+
 	return configs[configName];
 }
 
-let exportedConfigs = [];
 
-try {
-	exportedConfigs = main(process.argv);
-} catch (error) {
-	console.error(chalk.red(error.message));
-	process.exit(1);
-}
+export default (env) => {
+	let exportedConfigs = [];
 
-if (!exportedConfigs.length) {
-	// Nothing to do - for example where there are no external scripts to
-	// compile.
-	process.exit(0);
-}
+	try {
+		exportedConfigs = main(env);
+	} catch (error) {
+		console.error(error.message);
+		process.exit(1);
+	}
 
-module.exports = exportedConfigs;
+	if (!exportedConfigs.length) {
+		// Nothing to do - for example where there are no external scripts to
+		// compile.
+		process.exit(0);
+	}
+
+	return exportedConfigs;
+};
