@@ -15,7 +15,7 @@ import { Command } from './types';
  *
  * * [Main screen commands](https://github.com/laurent22/joplin/tree/dev/packages/app-desktop/gui/MainScreen/commands)
  * * [Global commands](https://github.com/laurent22/joplin/tree/dev/packages/app-desktop/commands)
- * * [Editor commands](https://github.com/laurent22/joplin/tree/dev/packages/app-desktop/gui/NoteEditor/commands/editorCommandDeclarations.ts)
+ * * [Editor commands](https://github.com/laurent22/joplin/tree/dev/packages/app-desktop/gui/NoteEditor/editorCommandDeclarations.ts)
  *
  * To view what arguments are supported, you can open any of these files
  * and look at the `execute()` command.
@@ -49,7 +49,7 @@ import { Command } from './types';
  *
  */
 export default class JoplinCommands {
-	/**
+    /**
      * <span class="platform-desktop">desktop</span> Executes the given
      * command.
      *
@@ -68,8 +68,8 @@ export default class JoplinCommands {
      * await joplin.commands.execute('newFolder', "SOME_FOLDER_ID");
      * ```
      */
-	execute(commandName: string, ...args: any[]): Promise<any | void>;
-	/**
+    execute(commandName: string, ...args: any[]): Promise<any | void>;
+    /**
      * <span class="platform-desktop">desktop</span> Registers a new command.
      *
      * ```typescript
@@ -85,5 +85,5 @@ export default class JoplinCommands {
      * });
      * ```
      */
-	register(command: Command): Promise<void>;
+    register(command: Command): Promise<void>;
 }
