@@ -5,6 +5,6 @@
  * so for now disable filters.
  */
 export default class JoplinFilters {
-	on(name: string, callback: Function): Promise<void>;
-	off(name: string, callback: Function): Promise<void>;
+    on(name: string, callback: Function): Promise<void>;
+    off(name: string, callback: Function): Promise<void>;
 }
