@@ -44,6 +44,12 @@ export default class JoplinViewsDialogs {
      */
     showMessageBox(message: string): Promise<number>;
     /**
+     * Displays a dialog to select a file or a directory. Same options and
+     * output as
+     * https://www.electronjs.org/docs/latest/api/dialog#dialogshowopendialogbrowserwindow-options
+     */
+    showOpenDialog(options: any): Promise<any>;
+    /**
      * Sets the dialog HTML content
      */
     setHtml(handle: ViewHandle, html: string): Promise<string>;
@@ -61,8 +67,7 @@ export default class JoplinViewsDialogs {
     open(handle: ViewHandle): Promise<DialogResult>;
     /**
      * Toggle on whether to fit the dialog size to the content or not.
-     * When set to false, the dialog stretches to fill the application
-     * window.
+     * When set to false, the dialog is set to 90vw and 80vh
      * @default true
      */
     setFitToContent(handle: ViewHandle, status: boolean): Promise<boolean>;
