@@ -1,17 +1,18 @@
+/* eslint indent: ["error", 2] */
 
 function plugin(CodeMirror) {
   CodeMirror.defineExtension('sortableApply', function(message) {
-    var sourceLine = message["source-line"];
-    var rows       = message["rows"];
-    var maxIdx     = Math.max(...rows);
-    if (maxIdx+1 != rows.length) { return; }
-    var sorted = this.getLine(sourceLine) + '\n' + this.getLine(sourceLine+1) + '\n';
-    for (var i=0; i<=maxIdx; i++){
-      sorted = sorted + this.getLine(sourceLine+2+rows[i]) + '\n';
+    const sourceLine = message['source-line'];
+    const rows = message['rows'];
+    const maxIdx = Math.max(...rows);
+    if (maxIdx + 1 !== rows.length) { return; }
+    let sorted = `${this.getLine(sourceLine)}\n${this.getLine(sourceLine + 1)}\n`;
+    for (let i = 0; i <= maxIdx; i++) {
+      sorted = `${sorted + this.getLine(sourceLine + 2 + rows[i])}\n`;
     }
-    this.setSelection(CodeMirror.Pos(sourceLine,0), CodeMirror.Pos(sourceLine + 3 + maxIdx, 0));
-    this.replaceSelection(sorted, "around");
-  })
+    this.setSelection(CodeMirror.Pos(sourceLine, 0), CodeMirror.Pos(sourceLine + 3 + maxIdx, 0));
+    this.replaceSelection(sorted, 'around');
+  });
 }
 
 module.exports = {
@@ -21,8 +22,8 @@ module.exports = {
       codeMirrorResources: [],
       codeMirrorOptions: {},
       assets: function() {
-        return [ ];
+        return [];
       },
-    }
-  }
-}
+    };
+  },
+};
